refactor: Table 관련 기능들 같은 패키지로 이동.
