+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.table.application;
 
 import java.util.Arrays;
 import java.util.List;
