+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,9 +1,9 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 import kitchenpos.order.dto.OrderTableRequest;
 
 public class TableGroupRequest {
