+++ src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.ui;
+package kitchenpos.table.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.TableService;
+import kitchenpos.table.application.TableService;
 import kitchenpos.order.dto.OrderTableRequest;
 import kitchenpos.order.dto.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
