+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -8,9 +8,10 @@ import java.util.Arrays;
 import java.util.Collections;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.table.dao.TableGroupDao;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.application.TableGroupService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
