+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -8,6 +8,7 @@ import java.util.Optional;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderTableDao;
 import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.application.TableService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
